Extracts PickingList and PickingListItem from Shipment class file to their own files
