+++ ecommerce/shipping/lib/shipping/events/shipping_address_added_to_shipment.rb
@@ -1,0 +1,9 @@
+module Shipping
+  class ShippingAddressAddedToShipment < Infra::Event
+    attribute :order_id, Infra::Types::UUID
+    attribute :line_1, Infra::Types::String
+    attribute :line_2, Infra::Types::String
+    attribute :line_3, Infra::Types::String
+    attribute :line_4, Infra::Types::String
+  end
+end
