+++ ecommerce/shipping/lib/shipping/picking_list_item.rb
@@ -1,0 +1,18 @@
+module Shipping
+  class PickingListItem
+    attr_reader :product_id, :quantity
+
+    def initialize(product_id)
+      @product_id = product_id
+      @quantity = 0
+    end
+
+    def increase
+      @quantity += 1
+    end
+
+    def decrease
+      @quantity -= 1
+    end
+  end
+end
